fix piped stdin and stdout, add possibility to print board, check pon movement.

+++ chess_referee/src/movement.rs
@@ -5,15 +5,15 @@ pub struct Movement {
 }
 
 impl Movement {
-    pub fn dir(&self) -> (u8,u8) {
-        (self.after.0 as u8 - self.before.0 as u8, self.after.1 as u8 - self.before.1 as u8)
+    pub fn dir(&self) -> (i8,i8) {
+        (self.after.0 as i8 - self.before.0 as i8, self.after.1 as i8 - self.before.1 as i8)
     }
 }
 
 impl From<[u8; 4]> for Movement {
     fn from(coord: [u8; 4]) -> Self {
-        let before = ((coord[0]-'a' as u8) as usize, (8-coord[1]) as usize);
-        let after = ((coord[2]-'a' as u8) as usize, (8-coord[3]) as usize);
+        let before = ((coord[0] as i8 -'a' as i8) as usize, (8-coord[1] as i8) as usize);
+        let after = ((coord[2] as i8 -'a' as i8) as usize, (8-coord[3] as i8) as usize);
         Movement { before, after }
     }
 }
